Made backend more clean
