+++ .vscode/launch.json
@@ -1,0 +1,15 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Launch with ts-node",
+      "program": "${workspaceFolder}/src/server.ts", // Adjust this path to your main entry file
+      "runtimeArgs": ["-r", "ts-node/register"],
+      "cwd": "${workspaceFolder}",
+      "sourceMaps": true, // Ensure source maps are enabled
+      "skipFiles": ["<node_internals>/**"]
+    }
+  ]
+}
