+++ package.json
@@ -13,6 +13,7 @@
   "description": "",
   "dependencies": {
     "@pothos/core": "^4.2.0",
+    "@pothos/plugin-errors": "^4.2.0",
     "@pothos/plugin-prisma": "^4.2.1",
     "@prisma/client": "^5.19.1",
     "@types/jsonwebtoken": "^9.0.6",
