+++ src/schema/index.ts
@@ -1,4 +1,3 @@
 import "./item";
 import "./user";
-import "./queries";
 import "./auth";
